feat
Made it so that watchface text color changes according to the background so that certain backgrounds have text colors that are visible. Currently the project is in perfect working condition, but it has a lot of log messages. The next commit will remove the log messages, but use this commit if you want to log what is going on